Implement session inactivity timeout in navbar
Added logic to automatically log out users after 30 minutes of inactivity by checking the last activity timestamp on each request. The session is destroyed and the user is redirected to the logout page if the inactivity threshold is exceeded. Also ensured session_start() is always called and added exit() after redirection for better flow control.